refactor: 프로젝트명 설정에 따른 디렉토리 수정
