--- src/main/java/poke/fromitive/attendance/AttendanceApplication.java
@@ -1,13 +1,0 @@
-package poke.fromitive.attendance;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.scheduling.annotation.EnableScheduling;
-
-@SpringBootApplication
-@EnableScheduling
-public class AttendanceApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(AttendanceApplication.class, args);
-    }
-}
