--- src/main/java/poke/fromitive/attendance/controller/AttendanceController.java
@@ -1,23 +1,0 @@
-package poke.fromitive.attendance.controller;
-
-import java.util.List;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-import poke.fromitive.attendance.response.AttendanceResponse;
-import poke.fromitive.attendance.service.AttendanceService;
-
-@RestController
-public class AttendanceController {
-    private final AttendanceService attendanceService;
-
-    public AttendanceController(final AttendanceService attendanceService) {
-        this.attendanceService = attendanceService;
-    }
-
-    @GetMapping("/crews")
-    ResponseEntity<List<AttendanceResponse>> getAttendance() {
-        List<AttendanceResponse> attendanceResponses = attendanceService.getNotAttendanceCrew();
-        return ResponseEntity.ok(attendanceResponses);
-    }
-}
