--- src/main/java/poke/fromitive/attendance/response/AttendanceResponse.java
@@ -1,9 +1,0 @@
-package poke.fromitive.attendance.response;
-
-import poke.fromitive.attendance.domain.Crew;
-
-public record AttendanceResponse(String name) {
-    public static AttendanceResponse fromCrew(Crew crew) {
-        return new AttendanceResponse(crew.getName());
-    }
-}
