+++ src/main/java/woowacourse/crewniverse/milkyway/AttendanceApplication.java
@@ -1,0 +1,13 @@
+package woowacourse.crewniverse.milkyway;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.scheduling.annotation.EnableScheduling;
+
+@SpringBootApplication
+@EnableScheduling
+public class AttendanceApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(AttendanceApplication.class, args);
+    }
+}
