+++ src/main/java/woowacourse/crewniverse/milkyway/httpclient/AttendanceCrawler.java
@@ -1,0 +1,9 @@
+package woowacourse.crewniverse.milkyway.httpclient;
+
+import java.io.IOException;
+import java.security.GeneralSecurityException;
+import woowacourse.crewniverse.milkyway.response.AttendanceSheetResponses;
+
+public interface AttendanceCrawler {
+    AttendanceSheetResponses execute() throws GeneralSecurityException, IOException;
+}
