+++ src/main/java/woowacourse/crewniverse/milkyway/response/AttendanceResponse.java
@@ -1,0 +1,9 @@
+package woowacourse.crewniverse.milkyway.response;
+
+import woowacourse.crewniverse.milkyway.domain.Crew;
+
+public record AttendanceResponse(String name) {
+    public static AttendanceResponse fromCrew(Crew crew) {
+        return new AttendanceResponse(crew.getName());
+    }
+}
