--- src/test/java/poke/fromitive/attendance/BaseSpringBootTest.java
@@ -1,16 +1,0 @@
-package poke.fromitive.attendance;
-
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.test.annotation.DirtiesContext;
-import org.springframework.test.context.ActiveProfiles;
-import poke.fromitive.attendance.config.TestCrawlerConfiguration;
-
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT, classes = {TestCrawlerConfiguration.class})
-@DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_CLASS)
-@ActiveProfiles("test")
-public abstract class BaseSpringBootTest {
-
-    @LocalServerPort
-    protected int port;
-}
