+++ src/test/java/woowacourse/crewniverse/milkyway/BaseSpringBootTest.java
@@ -1,0 +1,16 @@
+package woowacourse.crewniverse.milkyway;
+
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.web.server.LocalServerPort;
+import org.springframework.test.annotation.DirtiesContext;
+import org.springframework.test.context.ActiveProfiles;
+import woowacourse.crewniverse.milkyway.config.TestCrawlerConfiguration;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT, classes = {TestCrawlerConfiguration.class})
+@DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_CLASS)
+@ActiveProfiles("test")
+public abstract class BaseSpringBootTest {
+
+    @LocalServerPort
+    protected int port;
+}
