+++ src/test/java/woowacourse/crewniverse/milkyway/config/TestCrawlerConfiguration.java
@@ -1,0 +1,17 @@
+package woowacourse.crewniverse.milkyway.config;
+
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Primary;
+import woowacourse.crewniverse.milkyway.fake.FakeAttendanceCrawler;
+import woowacourse.crewniverse.milkyway.httpclient.AttendanceCrawler;
+
+@TestConfiguration
+public class TestCrawlerConfiguration {
+
+    @Bean
+    @Primary
+    AttendanceCrawler testAttendanceCrawler() {
+        return new FakeAttendanceCrawler();
+    }
+}
